Extracts ArmoryCatGrid & ArmoryCatDetailsCard from ArmoryCatsView
